Update README.md
In this code, we declare three variables: num1, num2, and sum. These variables will store the two numbers entered by the user and the sum of the numbers, respectively.

We use std::cout to prompt the user to enter the first number and std::cin to read the input and store it in the num1 variable.